Deleted yarn.lock file
I'm deleting the `yarn.lock` file because there haven't been any new dependencies installed in this commit. The `yarn.lock` file keeps track of the specific versions of the dependencies used in the project. Since there are no changes to the dependencies, there's no need to include any changes to the `yarn.lock` file in the main repository. Deleting it helps keep things organized and avoids unnecessary changes to the file.